Split __init__ and font functions into functions.py
